Moved Srt To SSA/ASS conversion to SubtitleUtils and made some minor changes.
