+++ src/main/java/net/pms/formats/v2/SubtitleUtils.java
@@ -18,13 +18,32 @@
  */
 package net.pms.formats.v2;
 
+import java.io.BufferedReader;
+import java.io.BufferedWriter;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileOutputStream;
+import java.io.FileWriter;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.io.OutputStreamWriter;
 import java.util.HashMap;
 import java.util.Map;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import net.pms.configuration.PmsConfiguration;
+import net.pms.dlna.DLNAMediaInfo;
 import net.pms.dlna.DLNAMediaSubtitle;
+import net.pms.io.OutputParams;
+import net.pms.util.StringUtil;
 import static org.apache.commons.lang.StringUtils.isBlank;
+import static org.apache.commons.lang.StringUtils.isNotBlank;
 import static org.mozilla.universalchardet.Constants.*;
 
 public class SubtitleUtils {
+	private static final Logger LOGGER = LoggerFactory.getLogger(SubtitleUtils.class);
 	private final static Map<String, String> fileCharsetToMencoderSubcpOptionMap = new HashMap<String, String>() {
 		private static final long serialVersionUID = 1L;
 
@@ -74,5 +93,117 @@ public class SubtitleUtils {
 			return null;
 		}
 		return fileCharsetToMencoderSubcpOptionMap.get(dlnaMediaSubtitle.getExternalFileCharacterSet());
+	}
+	
+	public static File ConvertSrtToAss(String SrtFile, double timeseek, PmsConfiguration configuration ) throws IOException {
+		File outputSubs = null;
+		outputSubs = new File(configuration.getTempFolder(), "FFmpeg" + System.currentTimeMillis() + ".ass");
+		BufferedReader input = new BufferedReader(new InputStreamReader(new FileInputStream(SrtFile), configuration.getSubtitlesCodepage()));
+		BufferedWriter output = new BufferedWriter(new OutputStreamWriter(new FileOutputStream(outputSubs)));
+		String line = null;
+		output.write("[Script Info]\n");
+		output.write("ScriptType: v4.00+\n");
+//			output.write("PlayResX: " + media.getWidth() + "\n"); // TODO Not clear how it works
+//			output.write("PlayResY: " + media.getHeight() + "\n");
+		output.write("\n");
+		output.write("[V4+ Styles]\n");
+		output.write("Format: Name, Fontname, Fontsize, PrimaryColour, SecondaryColour, OutlineColour, BackColour, Bold, Italic, Underline, BorderStyle, Outline, Shadow, Alignment, MarginL, MarginR, MarginV, AlphaLevel, Encoding\n");
+		StringBuilder s = new StringBuilder();
+		s.append("Style: Default,");
+
+		if (!configuration.getFont().isEmpty()) {
+			s.append(configuration.getFont()).append(",");
+		} else {
+			s.append("Arial,");
+		}
+
+		s.append( (int) 10 * Double.parseDouble(configuration.getMencoderAssScale())).append(","); // Fontsize
+
+		String primaryColour = Integer.toHexString(configuration.getSubsColor());
+		primaryColour = primaryColour.substring(6, 8) + primaryColour.substring(4, 6) + primaryColour.substring(2, 4); // Convert AARRGGBB format to BBGGRR
+
+		s.append("&H").append(primaryColour).append(","); // PrimaryColour
+		s.append("&Hffffff,"); 	// SecondaryColour
+		s.append("&H0,");		// OutlineColour
+		s.append("&H0,"); 		// BackColour
+		s.append("0,"); 		// Bold
+		s.append("0,"); 		// Italic
+		s.append("0,"); 		// Underline
+		s.append("1,"); 		// BorderStyle
+		s.append(configuration.getMencoderAssOutline()).append(","); // Outline
+		s.append(configuration.getMencoderAssShadow()).append(","); // Shadow
+		s.append("2,"); 		// Alignment
+		s.append("10,"); 		// MarginL
+		s.append("10,"); 		// MarginR
+		s.append("20,"); 		// MarginV
+		s.append("0,"); 		// AlphaLevel
+		s.append("0"); 			// Encoding
+		output.write(s.toString() + "\n");
+		output.write("\n");
+		output.write("[Events]\n");
+		output.write("Format: Layer, Start, End, Style, Text\n");
+			
+		String startTime;
+		String endTime;
+
+		while (( line = input.readLine()) != null) {
+			if (line .contains("-->")) {
+				startTime = line.substring(0, line.indexOf("-->") - 1).replaceAll(",", ".");
+				endTime = line.substring(line.indexOf("-->") + 4).replaceAll(",", ".");
+
+				// Apply time seeking
+				if (timeseek > 0) {
+					if (StringUtil.convertStringToTime(startTime) >= timeseek) {
+						startTime = StringUtil.convertTimeToString(StringUtil.convertStringToTime(startTime) - timeseek, false);
+						startTime = startTime.substring(1, startTime.length() - 1);
+						endTime = StringUtil.convertTimeToString(StringUtil.convertStringToTime(endTime) - timeseek, false);
+						endTime = endTime.substring(1, endTime.length() - 1);
+					} else {
+						continue;
+					}
+				}
+
+				s = new StringBuilder();
+				s.append("Dialogue: 0,");
+				s.append(startTime).append(",");
+				s.append(endTime).append(",");
+				s.append("Default").append(",");
+				s.append(convertTags(input.readLine())); 
+
+				if (isNotBlank(line = input.readLine())) {
+					s.append("\\N");
+					s.append(convertTags(line));
+				}
+
+				output.write(s.toString() + "\n");
+			}
+		}
+
+		input.close();
+		output.flush();
+		output.close();
+		outputSubs.deleteOnExit();
+		return outputSubs;
+
+	}
+
+	private static String convertTags(String text) {
+		 String tag = null;
+		 StringBuilder sb = new StringBuilder();
+		 String[] tmp = text.split("<");
+
+		 for (String s : tmp) {
+			 if (s.startsWith("/") && s.indexOf(">") == 2) {
+				 tag = s.substring(1, 2);
+				 sb.append("{\\").append(tag).append("0}").append(s.substring(3));
+			 } else if (s.indexOf(">") == 1) {
+				 tag = s.substring(0, 1);
+				 sb.append("{\\").append(tag).append("1}").append(s.substring(2));
+			 } else {
+				 sb.append(s);
+			 }
+		 }
+
+		 return sb.toString();
 	}
 }
